requirements.txt
streamlit==1.32.0
torch==2.2.1
torchvision==0.17.1
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.0
matplotlib==3.8.3
seaborn==0.13.0
pillow==10.1.0
joblib==1.3.2
tqdm==4.66.1
opencv-python==4.8.1